stop setting max log level
I don't remember why I did this here, but it was a bad idea.
As the log documentation writes "Libraries should avoid using the max
level features because they're global and can't be changed once they're
set." I was trying to see debug logs in an upstream package and couldn't
because it was overridden here.